consumer/requirements.txt
pandas
tweepy==3.10
pymongo
python-decouple
torch 
torchvision 
transformers
matplotlib
scipy
dnspython
kafka-python